Fix target vs host confusion (#37)
Since the build.rs is compiled and executed on the host architecture,
cfg!(target_ ...) in build.rs check for the host architecture, but not
the target architecture. If host and target architecture both use
gcc-ish platforms or MSVC, the checks based on cfg!(target_*) have the
desired effect. However, when one architecture accepts gcc flags, but
the other doesn't, builds break.

This issue was observed when trying to cross compile to
x86_64-pc-windows-gnu via cargo-cross.